requirements.txt
gpiozero
fastapi
uvicorn[standard]
jinja2
python-multipart

gradio_imageslider
gradio==4.29.0
matplotlib
opencv-python
torch
torchvision
ultralytics
pillow
numpy
transformers